backend/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.6
pymongo==4.4.1
python-dotenv==1.0.0